refactor: migrate from Vite+React SPA to Next.js App Router
- Remove old apps/web/ (Vite-based) and root app/ (stub Next.js)
- Create new Next.js App Router structure with file-based routing
- Migrate all 8 pages: Home, About, Login, Register, Session,
  Dashboard, Progress, Settings
- Migrate contexts (Auth, Emotion, Therapy) to root contexts/
- Add (protected) route group with auth-checking layout
- Extract SessionCanvas as dynamic import (SSR-disabled for Three.js)
- Add packages/core package.json with proper exports
- Update configs: tsconfig.json (paths), tailwind.config.js,
  postcss.config.js, package.json (deps cleanup)
- Update Dockerfile for Next.js standalone output
- Fix nginx.conf port mismatch (4000 → 8000)
- Update CLAUDE.md with new structure and WebRTC architecture

Key changes:
- react-router-dom → Next.js Link/useRouter/usePathname
- Vite build → Next.js build with standalone output
- Single framework, single build system
