+++ app/(protected)/layout.tsx
@@ -1,29 +1,43 @@
+'use client';
+
 import React, { useState } from 'react';
-import { Link, Outlet, useLocation, useNavigate } from 'react-router-dom';
-import { motion } from 'framer-motion';
-import { useAuth } from '../contexts/AuthContext';
+import Link from 'next/link';
+import { usePathname, useRouter } from 'next/navigation';
+import { useAuth } from '@/contexts/AuthContext';
 
-export const Layout: React.FC = () => {
+const menuItems = [
+  { path: '/dashboard', label: 'Dashboard', icon: '🏠' },
+  { path: '/session', label: 'New Session', icon: '🧠' },
+  { path: '/progress', label: 'Progress', icon: '📊' },
+  { path: '/settings', label: 'Settings', icon: '⚙️' }
+];
+
+export default function ProtectedLayout({ children }: { children: React.ReactNode }) {
   const [isSidebarOpen, setIsSidebarOpen] = useState(false);
-  const location = useLocation();
-  const navigate = useNavigate();
-  const { user, logout } = useAuth();
+  const pathname = usePathname();
+  const router = useRouter();
+  const { user, isAuthenticated, loading, logout } = useAuth();
 
-  const menuItems = [
-    { path: '/dashboard', label: 'Dashboard', icon: '🏠' },
-    { path: '/session', label: 'New Session', icon: '🧠' },
-    { path: '/progress', label: 'Progress', icon: '📊' },
-    { path: '/settings', label: 'Settings', icon: '⚙️' }
-  ];
+  if (loading) {
+    return (
+      <div className="min-h-screen bg-gradient-to-br from-purple-900 via-blue-900 to-indigo-900 flex items-center justify-center">
+        <div className="spinner"></div>
+      </div>
+    );
+  }
+
+  if (!isAuthenticated) {
+    router.push('/login');
+    return null;
+  }
 
   const handleLogout = () => {
     logout();
-    navigate('/');
+    router.push('/');
   };
 
   return (
     <div className="min-h-screen bg-gradient-to-br from-gray-900 via-purple-900 to-blue-900">
-      {/* Header */}
       <header className="bg-black/30 backdrop-blur-md border-b border-white/10">
         <div className="px-4 py-4">
           <div className="flex items-center justify-between">
@@ -36,23 +50,19 @@ export const Layout: React.FC = () => {
                   <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 6h16M4 12h16M4 18h16" />
                 </svg>
               </button>
-              <Link to="/dashboard" className="flex items-center space-x-2">
+              <Link href="/dashboard" className="flex items-center space-x-2">
                 <div className="w-10 h-10 bg-gradient-to-br from-blue-400 to-purple-600 rounded-lg flex items-center justify-center">
                   <span className="text-white text-xl">🧠</span>
                 </div>
                 <span className="text-xl font-bold text-white">EMDR-AI</span>
               </Link>
             </div>
-            
             <div className="flex items-center space-x-4">
               <div className="hidden md:flex items-center space-x-2 text-white/70">
                 <span>Welcome,</span>
                 <span className="font-semibold text-white">{user?.name || 'User'}</span>
               </div>
-              <button
-                onClick={handleLogout}
-                className="px-4 py-2 bg-white/10 text-white rounded-lg hover:bg-white/20 transition-colors"
-              >
+              <button onClick={handleLogout} className="px-4 py-2 bg-white/10 text-white rounded-lg hover:bg-white/20 transition-colors">
                 Logout
               </button>
             </div>
@@ -61,18 +71,15 @@ export const Layout: React.FC = () => {
       </header>
 
       <div className="flex">
-        {/* Sidebar */}
-        <aside className={`${
-          isSidebarOpen ? 'translate-x-0' : '-translate-x-full'
-        } lg:translate-x-0 fixed lg:static inset-y-0 left-0 z-50 w-64 bg-black/30 backdrop-blur-md border-r border-white/10 transition-transform duration-300`}>
-          <nav className="p-4 space-y-2">
+        <aside className={`${isSidebarOpen ? 'translate-x-0' : '-translate-x-full'} lg:translate-x-0 fixed lg:static inset-y-0 left-0 z-50 w-64 bg-black/30 backdrop-blur-md border-r border-white/10 transition-transform duration-300`}>
+          <nav className="p-4 space-y-2 mt-16 lg:mt-0">
             {menuItems.map(item => (
               <Link
                 key={item.path}
-                to={item.path}
+                href={item.path}
                 onClick={() => setIsSidebarOpen(false)}
                 className={`flex items-center space-x-3 px-4 py-3 rounded-xl transition-colors ${
-                  location.pathname === item.path
+                  pathname === item.path
                     ? 'bg-white/20 text-white'
                     : 'text-white/70 hover:bg-white/10 hover:text-white'
                 }`}
@@ -84,21 +91,14 @@ export const Layout: React.FC = () => {
           </nav>
         </aside>
 
-        {/* Main Content */}
         <main className="flex-1 p-4 lg:p-8">
-          <Outlet />
+          {children}
         </main>
       </div>
 
-      {/* Mobile sidebar overlay */}
       {isSidebarOpen && (
-        <div
-          className="lg:hidden fixed inset-0 bg-black/50 z-40"
-          onClick={() => setIsSidebarOpen(false)}
-        />
+        <div className="lg:hidden fixed inset-0 bg-black/50 z-40" onClick={() => setIsSidebarOpen(false)} />
       )}
     </div>
   );
-};
-
-export default Layout;
+}
