+++ app/login/page.tsx
@@ -1,22 +1,25 @@
+'use client';
+
 import React, { useState } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
+import Link from 'next/link';
+import { useRouter } from 'next/navigation';
 import { motion } from 'framer-motion';
 
-export const LoginPage: React.FC = () => {
+export default function LoginPage() {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
   const [rememberMe, setRememberMe] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
-  const navigate = useNavigate();
+  const router = useRouter();
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
     setIsLoading(true);
-    
-    // Simulate login
+
+    // TODO: Replace with real API call
     setTimeout(() => {
       setIsLoading(false);
-      navigate('/dashboard');
+      router.push('/dashboard');
     }, 1500);
   };
 
@@ -28,9 +31,8 @@ export const LoginPage: React.FC = () => {
         transition={{ duration: 0.5 }}
         className="w-full max-w-md"
       >
-        {/* Logo */}
         <div className="text-center mb-8">
-          <Link to="/" className="inline-block">
+          <Link href="/" className="inline-block">
             <div className="w-20 h-20 bg-gradient-to-br from-blue-400 to-purple-600 rounded-2xl flex items-center justify-center mx-auto mb-4">
               <span className="text-white text-4xl">🧠</span>
             </div>
@@ -38,15 +40,12 @@ export const LoginPage: React.FC = () => {
           </Link>
         </div>
 
-        {/* Login Form */}
         <div className="bg-white/10 backdrop-blur-md rounded-3xl p-8">
           <h2 className="text-2xl font-bold text-white mb-6 text-center">Welcome Back</h2>
-          
+
           <form onSubmit={handleSubmit} className="space-y-6">
             <div>
-              <label htmlFor="email" className="block text-white/70 text-sm mb-2">
-                Email Address
-              </label>
+              <label htmlFor="email" className="block text-white/70 text-sm mb-2">Email Address</label>
               <input
                 id="email"
                 type="email"
@@ -59,9 +58,7 @@ export const LoginPage: React.FC = () => {
             </div>
 
             <div>
-              <label htmlFor="password" className="block text-white/70 text-sm mb-2">
-                Password
-              </label>
+              <label htmlFor="password" className="block text-white/70 text-sm mb-2">Password</label>
               <input
                 id="password"
                 type="password"
@@ -75,15 +72,10 @@ export const LoginPage: React.FC = () => {
 
             <div className="flex items-center justify-between">
               <label className="flex items-center text-white/70 text-sm">
-                <input
-                  type="checkbox"
-                  checked={rememberMe}
-                  onChange={(e) => setRememberMe(e.target.checked)}
-                  className="mr-2"
-                />
+                <input type="checkbox" checked={rememberMe} onChange={(e) => setRememberMe(e.target.checked)} className="mr-2" />
                 Remember me
               </label>
-              <Link to="/forgot-password" className="text-blue-400 text-sm hover:text-blue-300">
+              <Link href="/forgot-password" className="text-blue-400 text-sm hover:text-blue-300">
                 Forgot password?
               </Link>
             </div>
@@ -109,33 +101,20 @@ export const LoginPage: React.FC = () => {
 
             <div className="mt-6 grid grid-cols-2 gap-3">
               <button className="flex items-center justify-center px-4 py-2 bg-white/10 border border-white/20 rounded-xl hover:bg-white/20 transition-colors">
-                <svg className="w-5 h-5 text-white" viewBox="0 0 24 24">
-                  <path fill="currentColor" d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"/>
-                  <path fill="currentColor" d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"/>
-                  <path fill="currentColor" d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"/>
-                  <path fill="currentColor" d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"/>
-                </svg>
-                <span className="ml-2 text-white/70">Google</span>
+                <span className="text-white/70">Google</span>
               </button>
               <button className="flex items-center justify-center px-4 py-2 bg-white/10 border border-white/20 rounded-xl hover:bg-white/20 transition-colors">
-                <svg className="w-5 h-5 text-white" fill="currentColor" viewBox="0 0 24 24">
-                  <path d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z"/>
-                </svg>
-                <span className="ml-2 text-white/70">GitHub</span>
+                <span className="text-white/70">GitHub</span>
               </button>
             </div>
           </div>
 
           <p className="mt-8 text-center text-white/60 text-sm">
-            Don't have an account?{' '}
-            <Link to="/register" className="text-blue-400 hover:text-blue-300">
-              Sign up for free
-            </Link>
+            Don&apos;t have an account?{' '}
+            <Link href="/register" className="text-blue-400 hover:text-blue-300">Sign up for free</Link>
           </p>
         </div>
       </motion.div>
     </div>
   );
-};
-
-export default LoginPage;
+}
