+++ app/register/page.tsx
@@ -1,8 +1,11 @@
+'use client';
+
 import React, { useState } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
+import Link from 'next/link';
+import { useRouter } from 'next/navigation';
 import { motion } from 'framer-motion';
 
-export const RegisterPage: React.FC = () => {
+export default function RegisterPage() {
   const [formData, setFormData] = useState({
     name: '',
     email: '',
@@ -13,7 +16,7 @@ export const RegisterPage: React.FC = () => {
   });
   const [isLoading, setIsLoading] = useState(false);
   const [errors, setErrors] = useState<Record<string, string>>({});
-  const navigate = useNavigate();
+  const router = useRouter();
 
   const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const { name, value, type, checked } = e.target;
@@ -21,7 +24,6 @@ export const RegisterPage: React.FC = () => {
       ...prev,
       [name]: type === 'checkbox' ? checked : value
     }));
-    // Clear error for this field
     if (errors[name]) {
       setErrors(prev => ({ ...prev, [name]: '' }));
     }
@@ -29,31 +31,25 @@ export const RegisterPage: React.FC = () => {
 
   const validate = () => {
     const newErrors: Record<string, string> = {};
-    
     if (!formData.name.trim()) newErrors.name = 'Name is required';
     if (!formData.email.trim()) newErrors.email = 'Email is required';
     if (!formData.password) newErrors.password = 'Password is required';
     if (formData.password.length < 8) newErrors.password = 'Password must be at least 8 characters';
-    if (formData.password !== formData.confirmPassword) {
-      newErrors.confirmPassword = 'Passwords do not match';
-    }
+    if (formData.password !== formData.confirmPassword) newErrors.confirmPassword = 'Passwords do not match';
     if (!formData.agreedToTerms) newErrors.terms = 'You must agree to the terms';
     if (!formData.agreedToAge) newErrors.age = 'You must be 18 or older';
-    
     setErrors(newErrors);
     return Object.keys(newErrors).length === 0;
   };
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
-    
     if (!validate()) return;
-    
     setIsLoading(true);
-    // Simulate registration
+    // TODO: Replace with real API call
     setTimeout(() => {
       setIsLoading(false);
-      navigate('/onboarding');
+      router.push('/dashboard');
     }, 1500);
   };
 
@@ -65,9 +61,8 @@ export const RegisterPage: React.FC = () => {
         transition={{ duration: 0.5 }}
         className="w-full max-w-md"
       >
-        {/* Logo */}
         <div className="text-center mb-8">
-          <Link to="/" className="inline-block">
+          <Link href="/" className="inline-block">
             <div className="w-20 h-20 bg-gradient-to-br from-blue-400 to-purple-600 rounded-2xl flex items-center justify-center mx-auto mb-4">
               <span className="text-white text-4xl">🧠</span>
             </div>
@@ -75,102 +70,44 @@ export const RegisterPage: React.FC = () => {
           </Link>
         </div>
 
-        {/* Registration Form */}
         <div className="bg-white/10 backdrop-blur-md rounded-3xl p-8">
           <h2 className="text-2xl font-bold text-white mb-6 text-center">Create Your Account</h2>
-          
+
           <form onSubmit={handleSubmit} className="space-y-5">
             <div>
-              <label htmlFor="name" className="block text-white/70 text-sm mb-2">
-                Full Name
-              </label>
-              <input
-                id="name"
-                name="name"
-                type="text"
-                value={formData.name}
-                onChange={handleChange}
-                className="w-full px-4 py-3 bg-white/10 border border-white/20 rounded-xl text-white placeholder-white/40 focus:outline-none focus:border-blue-400 transition-colors"
-                placeholder="John Doe"
-              />
+              <label htmlFor="name" className="block text-white/70 text-sm mb-2">Full Name</label>
+              <input id="name" name="name" type="text" value={formData.name} onChange={handleChange} className="w-full px-4 py-3 bg-white/10 border border-white/20 rounded-xl text-white placeholder-white/40 focus:outline-none focus:border-blue-400 transition-colors" placeholder="John Doe" />
               {errors.name && <p className="text-red-400 text-sm mt-1">{errors.name}</p>}
             </div>
 
             <div>
-              <label htmlFor="email" className="block text-white/70 text-sm mb-2">
-                Email Address
-              </label>
-              <input
-                id="email"
-                name="email"
-                type="email"
-                value={formData.email}
-                onChange={handleChange}
-                className="w-full px-4 py-3 bg-white/10 border border-white/20 rounded-xl text-white placeholder-white/40 focus:outline-none focus:border-blue-400 transition-colors"
-                placeholder="you@example.com"
-              />
+              <label htmlFor="email" className="block text-white/70 text-sm mb-2">Email Address</label>
+              <input id="email" name="email" type="email" value={formData.email} onChange={handleChange} className="w-full px-4 py-3 bg-white/10 border border-white/20 rounded-xl text-white placeholder-white/40 focus:outline-none focus:border-blue-400 transition-colors" placeholder="you@example.com" />
               {errors.email && <p className="text-red-400 text-sm mt-1">{errors.email}</p>}
             </div>
 
             <div>
-              <label htmlFor="password" className="block text-white/70 text-sm mb-2">
-                Password
-              </label>
-              <input
-                id="password"
-                name="password"
-                type="password"
-                value={formData.password}
-                onChange={handleChange}
-                className="w-full px-4 py-3 bg-white/10 border border-white/20 rounded-xl text-white placeholder-white/40 focus:outline-none focus:border-blue-400 transition-colors"
-                placeholder="••••••••"
-              />
+              <label htmlFor="password" className="block text-white/70 text-sm mb-2">Password</label>
+              <input id="password" name="password" type="password" value={formData.password} onChange={handleChange} className="w-full px-4 py-3 bg-white/10 border border-white/20 rounded-xl text-white placeholder-white/40 focus:outline-none focus:border-blue-400 transition-colors" placeholder="••••••••" />
               {errors.password && <p className="text-red-400 text-sm mt-1">{errors.password}</p>}
             </div>
 
             <div>
-              <label htmlFor="confirmPassword" className="block text-white/70 text-sm mb-2">
-                Confirm Password
-              </label>
-              <input
-                id="confirmPassword"
-                name="confirmPassword"
-                type="password"
-                value={formData.confirmPassword}
-                onChange={handleChange}
-                className="w-full px-4 py-3 bg-white/10 border border-white/20 rounded-xl text-white placeholder-white/40 focus:outline-none focus:border-blue-400 transition-colors"
-                placeholder="••••••••"
-              />
+              <label htmlFor="confirmPassword" className="block text-white/70 text-sm mb-2">Confirm Password</label>
+              <input id="confirmPassword" name="confirmPassword" type="password" value={formData.confirmPassword} onChange={handleChange} className="w-full px-4 py-3 bg-white/10 border border-white/20 rounded-xl text-white placeholder-white/40 focus:outline-none focus:border-blue-400 transition-colors" placeholder="••••••••" />
               {errors.confirmPassword && <p className="text-red-400 text-sm mt-1">{errors.confirmPassword}</p>}
             </div>
 
             <div className="space-y-3">
               <label className="flex items-start text-white/70 text-sm">
-                <input
-                  name="agreedToAge"
-                  type="checkbox"
-                  checked={formData.agreedToAge}
-                  onChange={handleChange}
-                  className="mr-2 mt-1"
-                />
+                <input name="agreedToAge" type="checkbox" checked={formData.agreedToAge} onChange={handleChange} className="mr-2 mt-1" />
                 <span>I confirm that I am 18 years or older</span>
               </label>
               {errors.age && <p className="text-red-400 text-sm ml-6">{errors.age}</p>}
 
               <label className="flex items-start text-white/70 text-sm">
-                <input
-                  name="agreedToTerms"
-                  type="checkbox"
-                  checked={formData.agreedToTerms}
-                  onChange={handleChange}
-                  className="mr-2 mt-1"
-                />
-                <span>
-                  I agree to the{' '}
-                  <Link to="/terms" className="text-blue-400 hover:text-blue-300">Terms of Service</Link>
-                  {' '}and{' '}
-                  <Link to="/privacy" className="text-blue-400 hover:text-blue-300">Privacy Policy</Link>
-                </span>
+                <input name="agreedToTerms" type="checkbox" checked={formData.agreedToTerms} onChange={handleChange} className="mr-2 mt-1" />
+                <span>I agree to the Terms of Service and Privacy Policy</span>
               </label>
               {errors.terms && <p className="text-red-400 text-sm ml-6">{errors.terms}</p>}
             </div>
@@ -186,22 +123,16 @@ export const RegisterPage: React.FC = () => {
 
           <div className="mt-6 p-4 bg-yellow-500/20 rounded-xl border border-yellow-500/30">
             <p className="text-yellow-200 text-sm">
-              ⚠️ <strong>Important:</strong> EMDR-AI is a wellness tool and not a replacement for 
-              professional mental health care. If you're experiencing severe symptoms or crisis, 
-              please seek help from a qualified professional.
+              <strong>Important:</strong> EMDR-AI is a wellness tool and not a replacement for professional mental health care.
             </p>
           </div>
 
           <p className="mt-6 text-center text-white/60 text-sm">
             Already have an account?{' '}
-            <Link to="/login" className="text-blue-400 hover:text-blue-300">
-              Sign in
-            </Link>
+            <Link href="/login" className="text-blue-400 hover:text-blue-300">Sign in</Link>
           </p>
         </div>
       </motion.div>
     </div>
   );
-};
-
-export default RegisterPage;
+}
