--- apps/web/.env.example
@@ -1,14 +1,0 @@
-# MorphCast SDK License Key (required for emotion recognition)
-VITE_MORPHCAST_LICENSE_KEY=your_license_key_here
-
-# API Configuration
-VITE_API_URL=http://localhost:4000/api
-VITE_WS_URL=ws://localhost:4000
-
-# Optional: Analytics
-VITE_GA_TRACKING_ID=
-VITE_SENTRY_DSN=
-
-# Optional: Third-party services
-VITE_STRIPE_PUBLIC_KEY=
-VITE_CLOUDINARY_CLOUD_NAME=
