--- apps/web/src/App.tsx
@@ -1,47 +1,0 @@
-import React from 'react';
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import { AnimatePresence } from 'framer-motion';
-
-// Pages
-import HomePage from './pages/HomePage';
-import AboutPage from './pages/AboutPage';
-import SessionPage from './pages/SessionPage';
-import ProgressPage from './pages/ProgressPage';
-import LoginPage from './pages/LoginPage';
-import RegisterPage from './pages/RegisterPage';
-import DashboardPage from './pages/DashboardPage';
-import SettingsPage from './pages/SettingsPage';
-
-// Components
-import ProtectedRoute from './components/ProtectedRoute';
-import Layout from './components/Layout';
-
-function App() {
-  return (
-    <Router>
-      <AnimatePresence mode="wait">
-        <Routes>
-          {/* Public Routes */}
-          <Route path="/" element={<HomePage />} />
-          <Route path="/about" element={<AboutPage />} />
-          <Route path="/login" element={<LoginPage />} />
-          <Route path="/register" element={<RegisterPage />} />
-          
-          {/* Semi-Protected Routes (can be used without login but with limited features) */}
-          <Route path="/session" element={<SessionPage />} />
-          
-          {/* Protected Routes */}
-          <Route element={<ProtectedRoute />}>
-            <Route element={<Layout />}>
-              <Route path="/dashboard" element={<DashboardPage />} />
-              <Route path="/progress" element={<ProgressPage />} />
-              <Route path="/settings" element={<SettingsPage />} />
-            </Route>
-          </Route>
-        </Routes>
-      </AnimatePresence>
-    </Router>
-  );
-}
-
-export default App;
