+++ nginx.conf
@@ -37,7 +37,7 @@ http {
 
         # API proxy (if needed)
         location /api {
-            proxy_pass http://backend:4000;
+            proxy_pass http://backend:8000;
             proxy_http_version 1.1;
             proxy_set_header Upgrade $http_upgrade;
             proxy_set_header Connection 'upgrade';
