+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "emdr-ai-therapy-assistant",
   "version": "0.1.0",
-  "description": "Революционная платформа виртуальной терапии с использованием ИИ и распознавания эмоций",
+  "description": "Revolutionary virtual therapy platform combining EMDR techniques with AI-powered emotion recognition",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -15,19 +15,13 @@
     "format": "prettier --write ."
   },
   "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@mui/material": "^5.14.0",
-    "@mui/icons-material": "^5.14.0",
-    "@reduxjs/toolkit": "^1.9.5",
-    "@tensorflow/tfjs": "^4.10.0",
+    "@react-three/drei": "^9.88.0",
+    "@react-three/fiber": "^8.13.0",
     "framer-motion": "^10.16.0",
     "next": "^13.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.1.0",
     "three": "^0.155.0",
-    "@react-three/fiber": "^8.13.0",
     "tone": "^14.7.77",
     "socket.io-client": "^4.7.0"
   },
@@ -38,29 +32,22 @@
     "@types/three": "^0.155.0",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
+    "autoprefixer": "^10.4.16",
     "eslint": "^8.45.0",
     "eslint-config-next": "^13.4.0",
     "eslint-config-prettier": "^8.8.0",
-    "prettier": "^3.0.0",
-    "typescript": "^5.1.0",
     "jest": "^29.6.0",
     "@testing-library/react": "^13.4.0",
-    "@testing-library/jest-dom": "^5.17.0"
+    "@testing-library/jest-dom": "^5.17.0",
+    "postcss": "^8.4.31",
+    "prettier": "^3.0.0",
+    "tailwindcss": "^3.3.5",
+    "typescript": "^5.1.0"
   },
   "engines": {
     "node": ">=18.0.0",
     "npm": ">=9.0.0"
   },
-  "keywords": [
-    "emdr",
-    "therapy",
-    "ai",
-    "emotion-recognition",
-    "mental-health",
-    "psychotherapy",
-    "trauma",
-    "healing"
-  ],
   "author": "EMDR-AI Team",
   "license": "MIT",
   "repository": {
