+++ packages/core/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@emdr42/core",
+  "version": "0.1.0",
+  "private": true,
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts",
+    "./patterns/movements": "./src/patterns/movements.ts",
+    "./services/emotion-recognition": "./src/services/emotion-recognition.ts",
+    "./services/audio-asmr": "./src/services/audio-asmr.ts"
+  }
+}
