+++ packages/core/src/index.ts
@@ -1,0 +1,3 @@
+export { MovementPatterns } from './patterns/movements';
+export { EmotionRecognitionService } from './services/emotion-recognition';
+export { AudioASMRService } from './services/audio-asmr';
